Clean up terra.sh to handle console sub-command

+++ _infra/acceptance/terra.sh
@@ -2,13 +2,13 @@
 
 # Helper script for using terraform commands in a workspace.
 # Don't use this script if doing anything other then mundane main commands like
-# plan, apply, or destroy.
+# console, plan, apply, or destroy.
 
 set -o nounset
 set -o pipefail
 set -o errexit
 
-# Should be plan, apply, or destroy
+# Should be console, plan, apply, or destroy
 terraform_command=$1
 
 test $(basename $PWD) = "_infra" || (echo "Must run this script from the _infra directory." && exit 1)
@@ -17,9 +17,10 @@ test $(basename $PWD) = "_infra" || (echo "Must run this script from the _infra 
 script_dir=$(dirname $(realpath $0))
 workspace=$(basename $script_dir)
 project_dir=$(dirname $PWD)
+
 project_version=$(jq -r '.version' $project_dir/package.json)
 
-artifact_dist_tar_gz=puzzle-massive-$(jq -r '.version' ../package.json).tar.gz
+artifact_dist_tar_gz=puzzle-massive-$project_version.tar.gz
 test -e $project_dir/$artifact_dist_tar_gz || (echo "Must create a versioned artifact file at path $project_dir/$artifact_dist_tar_gz before running this command." && exit 1)
 
 project_description="Temporary instance for acceptance"
@@ -30,8 +31,11 @@ echo "Using versioned artifact dist file: '$project_dir/$artifact_dist_tar_gz'"
 
 set -x
 
-rm -f $script_dir/puzzle-massive-*.tar.gz
-cp $project_dir/$artifact_dist_tar_gz $script_dir/
+existing_artifact="$(echo $script_dir/puzzle-massive-*.tar.gz)"
+if [ ! -e "$existing_artifact" -o "$(md5sum $existing_artifact | cut -f1 -d ' ')" != "$(md5sum $project_dir/$artifact_dist_tar_gz | cut -f1 -d ' ')" ]; then
+  rm -f $script_dir/puzzle-massive-*.tar.gz
+  cp $project_dir/$artifact_dist_tar_gz $script_dir/
+fi
 
 terraform workspace select $workspace || \
   terraform workspace new $workspace
